requirements.txt
Django==4.2.0
djangorestframework==3.14.0
python-dateutil==2.8.2
pandas==2.0.2
python-magic==0.4.27